Fix unexpected tables being created
Cause: CTM was applied to each TextElement in a TextObject, thus messing
up its Tx and Ty. When TableDetector sees that Ty is not 0(indicating next
TextElement is in the same line), it sees the TextObject as multi-linear,
hence, it and other TextObjects in the same line are flagged as columns
of a table.

Fix: CTM must therefore be only applied to first TextElement.